refactor: 테스트 분리 및 가독성을 위한 메소드 명 수정
