+++ src/test/java/roomescape/PageTest.java
@@ -1,0 +1,18 @@
+package roomescape;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+public class PageTest {
+	@Test
+	void page() {
+		RestAssured.given().log().all()
+				.when().get("/")
+				.then().log().all()
+				.statusCode(200);
+	}
+}
